feat:차트 탭 분리
