+++ src/pages/LiveChartPage/components/PriceChartInfo.jsx
@@ -66,7 +66,7 @@ const PriceChartInfo = () => {
       </div>
 
       {/* 시세 범위 선택 */}
-      <div className="mt-5 flex justify-center items-center gap-4 sm:gap-4">
+      <div className="mt-5 flex justify-center items-center gap-4">
         {[
           { label: '실시간 시세', value: 'realtime' },
           { label: '최근 시세', value: 'today' },
@@ -74,7 +74,7 @@ const PriceChartInfo = () => {
         ].map(({ label, value }, index, arr) => (
           <React.Fragment key={value}>
             <span
-              className={`text-[13px] sm:text-[14px] cursor-pointer ${
+              className={`text-[13px] sm:text-[14px] cursor-pointer  ${
                 selectedRange === value ? 'text-[#EB008B] font-bold' : 'text-[#2C2C2C]'
               }`}
               onClick={() => setSelectedRange(value)}
